# docs/swagger.yaml
basePath: /
definitions:
...
  auth.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        example: bonheur@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      token:
        example: JWT_TOKEN
        type: string
      user:
        $ref: '#/definitions/auth.User'
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: bonheur@example.com
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      username:
        example: bonheur
        type: string
    required:
    - email
    - password
    - username
    type: object
  auth.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/auth.User'
    type: object
  auth.User:
    properties:
      email:
        type: string
      id:
        format: int64
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
  betting.AutoInput:
    properties:
      amount:
        type: string
      auto_cashout_multiplier:
        type: string
      enabled:
        type: boolean
    type: object
  betting.AutoSetting:
    properties:
      amount:
        type: string
      auto_cashout_multiplier:
        type: string
      bet_number:
        type: integer
      enabled:
        type: boolean
      last_error:
        type: string
      last_round_id:
        type: integer
    type: object
  betting.CancelResult:
    properties:
      bet_id:
        type: integer
      refunded_amount:
        type: string
      remaining_balance:
        type: string
      status:
        type: string
    type: object
  betting.PlaceBetRequest:
    properties:
      amount:
        type: string
      auto_cashout_multiplier:
        type: string
      bet_number:
        type: integer
      round_id:
        type: integer
    type: object
  cashout.CashoutResponse:
    properties:
      bet_amount:
        type: number
      bet_id:
        type: integer
      multiplier:
        type: number
      payout:
        type: number
      remaining_balance:
        type: number
    type: object
  deposit.CreateRequest:
    properties:
      amount:
        type: string
      provider:
        type: string
    type: object
  deposit.Deposit:
    properties:
      amount:
        type: number
      completed_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      provider:
        type: string
      provider_reference:
        type: string
      status:
        $ref: '#/definitions/deposit.Status'
      user_id:
        type: integer
    type: object
  deposit.Status:
    enum:
    - PENDING
    - PROCESSING
    - COMPLETED
    - FAILED
    - CANCELLED
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusProcessing
    - StatusCompleted
    - StatusFailed
    - StatusCancelled
  game.GameRound:
    properties:
      betting_closes_at:
        type: string
      betting_opened_at:
        type: string
      client_seed:
        type: string
      crash_point:
        type: number
      created_at:
        type: string
      ended_at:
        type: string
      house_edge:
        type: number
      id:
        type: integer
      nonce:
        type: integer
      round_number:
        type: integer
      server_seed:
        type: string
      server_seed_hash:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/game.RoundStatus'
    type: object
  game.RoundStatus:
    enum:
    - CREATED
    - BETTING_OPEN
    - BETTING_CLOSED
    - RUNNING
    - CRASHED
    - SETTLED
    type: string
    x-enum-varnames:
    - RoundCreated
    - RoundBettingOpen
    - RoundBettingClosed
    - RoundRunning
    - RoundCrashed
    - RoundSettled
  game.Snapshot:
    properties:
      current_multiplier:
        type: string
      phase:
        type: string
      running:
        $ref: '#/definitions/game.GameRound'
      seconds_remaining:
        type: integer
      server_time:
        type: string
      upcoming:
        $ref: '#/definitions/game.GameRound'
    type: object
  game.Verification:
    properties:
      message:
        type: string
      verified:
        type: boolean
    type: object
  history.PlayerBet:
    properties:
      amount:
        type: string
      auto_cashout_multiplier:
        type: string
      bet_number:
        type: integer
      can_cancel:
        type: boolean
      can_cashout:
        type: boolean
      cashout_multiplier:
        type: string
      id:
        type: integer
      payout:
        type: string
      potential_payout:
        type: string
      round_id:
        type: integer
      round_number:
        type: integer
      round_status:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
...
  wallet.BalanceResponse:
    properties:
      balance:
        example: "1000.00"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  withdrawal.CreateRequest:
    properties:
      amount:
        type: string
      provider:
        type: string
    type: object
  withdrawal.CreateResult:
    properties:
      remaining_balance:
        type: number
      withdrawal:
        $ref: '#/definitions/withdrawal.Withdrawal'
    type: object
  withdrawal.Status:
    enum:
    - PENDING
    - PROCESSING
    - COMPLETED
    - FAILED
    - CANCELLED
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusProcessing
    - StatusCompleted
    - StatusFailed
    - StatusCancelled
  withdrawal.Withdrawal:
    properties:
      amount:
        type: number
      completed_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      provider:
        type: string
      provider_reference:
        type: string
      status:
        $ref: '#/definitions/withdrawal.Status'
      user_id:
        type: integer
    type: object
host: localhost:7000
info:
  contact: {}
  description: HTTP API for the Aviator real-money game backend.
  title: Aviator Backend API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with an email and password and returns a JWT.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Log in
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a PLAYER account. The submitted role is never accepted;
        username and email are required, and the password must be at least 8 characters.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Register a user
      tags:
      - auth
  /api/bets:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/history.PlayerBet'
            type: array
      security:
      - BearerAuth: []
      summary: List own bets (latest 200)
      tags:
      - betting
    post:
      consumes:
      - application/json
      description: Places one of the user's bets in an open game round.
      parameters:
      - description: Bet details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/betting.PlaceBetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or bet
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Place a bet
      tags:
      - betting
  /api/bets/{id}/cancel:
    post:
      parameters:
      - description: Bet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/betting.CancelResult'
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel an active bet while betting is open
      tags:
      - betting
  /api/bets/{id}/cashout:
    post:
      consumes:
      - application/json
      description: Cashes out an active bet using the current server-side multiplier.
      parameters:
      - description: Bet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cashout.CashoutResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cash out a bet
      tags:
      - cashout
  /api/bets/auto:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/betting.AutoSetting'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get or update own automatic bet settings
      tags:
      - betting
  /api/bets/auto/{panel}:
    put:
      consumes:
      - application/json
      description: Automatic bets continue without a browser. One attempt per round
        and panel. A failed bet disables the setting. Auto cashout is captured on
        each accepted bet.
      parameters:
      - description: Panel 1 or 2
        in: path
        name: panel
        required: true
        type: integer
      - description: Settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/betting.AutoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/betting.AutoSetting'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update own server-side auto bet settings
      tags:
      - betting
  /api/deposits:
    get:
      description: Returns deposits for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/deposit.Deposit'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to load deposits
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List deposits
      tags:
      - deposits
    post:
      consumes:
      - application/json
      description: Creates a deposit request for the authenticated user.
      parameters:
      - description: Deposit details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/deposit.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/deposit.Deposit'
        "400":
          description: Invalid request or deposit
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a deposit
      tags:
      - deposits
...
  /api/game/rounds/{id}:
    get:
      description: Returns a game round by its database ID.
      parameters:
      - description: Round ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.GameRound'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get game round
      tags:
      - game
  /api/game/rounds/{id}/fairness:
    get:
      parameters:
      - description: Round ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.GameRound'
      summary: Get round fairness commitment or completed reveal
      tags:
      - game
  /api/game/rounds/{id}/verify:
    get:
      parameters:
      - description: Round ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.Verification'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify a settled round's seed commitment and crash result
      tags:
      - game
  /api/game/rounds/current:
    get:
      description: Returns running and upcoming rounds, server time and live multiplier;
        future secrets are redacted.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.Snapshot'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current game round
      tags:
      - game
...
  /api/wallet/balance:
    get:
      description: Returns the current wallet balance for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.BalanceResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get wallet balance
      tags:
      - wallet
  /api/wallet/transactions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List own wallet transactions (latest 200)
      tags:
      - wallet
  /api/withdrawals:
    get:
      description: Returns withdrawals for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/withdrawal.Withdrawal'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to load withdrawals
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List withdrawals
      tags:
      - withdrawals
    post:
      consumes:
      - application/json
      description: Creates a withdrawal request for the authenticated user.
      parameters:
      - description: Withdrawal details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/withdrawal.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/withdrawal.CreateResult'
        "400":
          description: Invalid request or withdrawal
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "409":
          description: Insufficient balance
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a withdrawal
      tags:
      - withdrawals
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"